Move general functions out of SyrTransliterator into SyrTools for other uses
